LPS-75781 Move interfaces into ***.internal.api
