+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/main/java/com/liferay/portal/configuration/extender/PropertiesFileNamedConfigurationContent.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.configuration.extender;
 
+import com.liferay.portal.configuration.extender.internal.api.NamedConfigurationContent;
+
 import java.io.IOException;
 import java.io.InputStream;
