+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/main/java/com/liferay/portal/configuration/extender/internal/ConfiguratorExtender.java
@@ -15,10 +15,10 @@
 package com.liferay.portal.configuration.extender.internal;
 
 import com.liferay.osgi.felix.util.AbstractExtender;
-import com.liferay.portal.configuration.extender.BundleStorage;
-import com.liferay.portal.configuration.extender.ConfigurationDescriptionFactory;
-import com.liferay.portal.configuration.extender.NamedConfigurationContent;
-import com.liferay.portal.configuration.extender.NamedConfigurationContentFactory;
+import com.liferay.portal.configuration.extender.internal.api.BundleStorage;
+import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescriptionFactory;
+import com.liferay.portal.configuration.extender.internal.api.NamedConfigurationContent;
+import com.liferay.portal.configuration.extender.internal.api.NamedConfigurationContentFactory;
 import com.liferay.portal.kernel.util.StringBundler;
 
 import java.io.IOException;
