+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/test/java/com/liferay/portal/configuration/extender/internal/ConfigurationDescriptionFactoryImplTest.java
@@ -14,12 +14,12 @@
 
 package com.liferay.portal.configuration.extender.internal;
 
-import com.liferay.portal.configuration.extender.ConfigurationDescription;
-import com.liferay.portal.configuration.extender.ConfigurationDescriptionFactory;
 import com.liferay.portal.configuration.extender.FactoryConfigurationDescription;
-import com.liferay.portal.configuration.extender.NamedConfigurationContent;
 import com.liferay.portal.configuration.extender.PropertiesFileNamedConfigurationContent;
 import com.liferay.portal.configuration.extender.SingleConfigurationDescription;
+import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescription;
+import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescriptionFactory;
+import com.liferay.portal.configuration.extender.internal.api.NamedConfigurationContent;
 import com.liferay.portal.kernel.util.Supplier;
 
 import java.io.ByteArrayInputStream;
